Rename shipping-details to addresses
