+++ src/themes/icmaa-imp/resource/i18n/de-DE.csv
@@ -149,7 +149,7 @@
 "My shipping details","Meine Versanddetails"
 "Name must have at least 3 letters.","Ihr Name muss aus mindestens 3 Zeichen bestehen."
 "New Luma Yoga Collection","Die neue Luma Yoga Kollektion"
-"New password *","Neues Passwort *"
+"New password","Neues Passwort"
 "Newsletter","Newsletter"
 "No orders yet","Noch keine Bestellungen"
 "No products found!","Keine Produkte gefunden!"
@@ -170,6 +170,7 @@
 "Order informations","Bestellinformationen"
 "Orders","Bestellungen"
 "Our {department}","Unsere {department}"
+"Password has successfully been changed","Das Passwort wurde erfolgreich geändert"
 "Password must have at least 8 letters.","Das Passwort muss mindestens 8 Zeichen enthalten."
 "Password","Passwort"
 "Passwords must be identical","Die Passwörter müssen identisch sein"
@@ -210,7 +211,7 @@
 "Remember me","Eingeloggt bleiben"
 "Remove from compare","Von Vergleichsliste entfernen"
 "Remove","Entfernen"
-"Repeat new password *","Neues Passwort wiederholen *"
+"Repeat new password","Neues Passwort wiederholen"
 "Repeat password","Passwort wiederholen"
 "Request size","Größe anfordern"
 "Reset password","Passwort zurücksetzen"
