+++ src/themes/icmaa-imp/resource/i18n/en-US.csv
@@ -173,6 +173,7 @@
 "Order informations","Order informations"
 "Orders","Orders"
 "Our {department}","Our {department}"
+"Password has successfully been changed","Password has successfully been changed"
 "Password must have at least 8 letters.","Password must have at least 8 letters."
 "Password","Password"
 "Passwords must be identical","Passwords must be identical"
